Development

+++ assets/scripts/engine/events.js
@@ -11,7 +11,7 @@ const addHandlers = () => {
   // $(document).on( 'keydown click', playSound)
   $('#library').on('click', '.set-collection', onSetCollection)
   $('#library').on('click', '#update-button', onUpdate)
-  $('#update-collection').on('submit', onUpdateCollection)
+  // $('#update-collection').on('submit', onUpdateCollection)
   $('#library').on('click', '.delete-collection', onDeleteCollection)
   $('#create-collection').on('submit', onCreateCollection)
   $('#get_all_collections').on('click', onShowAllCollections)
@@ -19,7 +19,6 @@ const addHandlers = () => {
 
 const onCreateCollection = function (event) {
   event.preventDefault()
-  console.log('token is' + store.user.token)
   const data = getFormFields(event.target)
   let sounds = []
   for (let i = 0; i < 16; i++) {
@@ -35,9 +34,15 @@ const onCreateCollection = function (event) {
 const onDeleteCollection = function (event) {
   event.preventDefault()
   const collectionId = $(event.target).closest('section').data('id')
-  api.deleteCollection(collectionId)
-    .then(ui.DeleteCollectionSuccess)
-    .catch(ui.failure)
+  const index = store.collections.findIndex(p => p.id === collectionId)
+  if ((store.collections[index].user !== null) && (store.collections[index].user.id === store.user.id)) {
+    api.deleteCollection(collectionId)
+      .then(ui.DeleteCollectionSuccess)
+      .catch(ui.failure)
+  } else {
+    $('#browse-message').text('You can`t change a collection that you didn`t create!')
+    $('#browse-message').css('color', 'red')
+  }
 }
 
 const onSetCollection = function (event) {
@@ -47,11 +52,19 @@ const onSetCollection = function (event) {
     .then(ui.setCollectionSuccess)
     .catch(ui.failure)
 }
-
-const onUpdate = function () {
-  $('#browse-collection').hide()
+const onUpdate = function (event) {
   const collectionId = $(event.target).closest('section').data('id')
   store.collectionId = collectionId
+  const index = store.collections.findIndex(p => p.id === collectionId)
+  if (store.collections[index].user !== null) {
+    if (store.collections[index].user.id === store.user.id) {
+      $('#browse-collection').modal('hide')
+      $('#general-message').text('')
+      $('#update').modal('show')
+    } else {
+      $('#browse-message').text('You can`t change a collection that you didn`t create!')
+    }
+  }
 }
 
 const onUpdateCollection = function (event) {
@@ -61,7 +74,6 @@ const onUpdateCollection = function (event) {
   for (let i = 0; i < 16; i++) {
     sounds.push(data[i])
   }
-  console.log('sounds are: ' + sounds)
   sounds = ',' + sounds.toString() + ','
   data.collection.sounds = sounds
   api.updateCollection(data)
@@ -70,15 +82,14 @@ const onUpdateCollection = function (event) {
 }
 
 const playSound = function (event) {
-  const key = event.which
-  const sound = new Audio()
-  console.log(store.pad)
-  console.log(key)
-  const index = myPadKeys.indexOf(key)
-  console.log('index is :' + index)
-  $('#' + key).css('background-color', 'red')
-  sound.src = store.pad[index]
-  sound.play()
+  if (store.padIsReady === true) {
+    const key = event.which
+    const sound = new Audio()
+    const index = myPadKeys.indexOf(key)
+    sound.src = store.pad[index]
+    sound.play()
+    $('#general-message').text('Sound Played')
+  }
 }
 
 const onShowAllCollections = function () {
